Add Misc slides section inside Announcement overlay
