+++ www/locales/en.json
@@ -94,6 +94,7 @@
     "ADD_ANNOUNCEMENT_SLIDE": "Add Announcement Slide",
     "ADD_ANNOUNCEMENT_TEXT": "Announcement goes here...",
     "ADD_ANNOUNCEMENT_TEXT_GURMUKHI": "GoSxw ie`Qy ilKo...",
+    "ADD_SLIDES": "Add Slides",
     "ADD_DHAN_GURU": "Add Dhan Guru",
     "ADD_EMPTY_SLIDE": "Add Empty Slide",
     "ADD_WAHEGURU_SLIDE": "Add Waheguru Slide",
