+++ www/main/common/hooks/index.js
@@ -1,1 +1,2 @@
 export { useKeys } from './useKeys';
+export { useSlides } from './useSlides';
